refactor: split up orderProgressBar mixed content
