+++ apps/cowswap-frontend/src/modules/application/containers/App/Updaters.tsx
@@ -18,7 +18,7 @@ import { BalancesCombinedUpdater } from 'modules/combinedBalances/updater/Balanc
 import { InFlightOrderFinalizeUpdater } from 'modules/ethFlow'
 import { CowEventsUpdater, InjectedWidgetUpdater, useInjectedWidgetParams } from 'modules/injectedWidget'
 import { FinalizeTxUpdater } from 'modules/onchainTransactions'
-import { ProgressBarExecutingOrdersUpdater } from 'modules/orderProgressBarMixed/orderProgressBar'
+import { ProgressBarExecutingOrdersUpdater } from 'modules/orderProgressBar'
 import { OrdersNotificationsUpdater } from 'modules/orders'
 import { useOnTokenListAddingError, useSourceChainId } from 'modules/tokensList'
 import { TradeType, useTradeTypeInfo } from 'modules/trade'
