+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/steps/FinishedStep.tsx
@@ -17,14 +17,13 @@ import { AMM_LOGOS } from 'legacy/components/AMMsLogo'
 import { Order } from 'legacy/state/orders/actions'
 import { useIsDarkMode } from 'legacy/state/user/hooks'
 
-import { OrderProgressBarStepName, SolverCompetition } from 'modules/orderProgressBarMixed/orderProgressBar'
-
 import { CowSwapAnalyticsCategory, toCowSwapGtmEvent } from 'common/analytics/types'
 import { SurplusData } from 'common/hooks/useGetSurplusFiatValue'
 import { getIsCustomRecipient } from 'utils/orderUtils/getIsCustomRecipient'
 
 import * as styledEl from './styled'
 
+import { OrderProgressBarStepName, SolverCompetition } from '../../types'
 import { CHAIN_SPECIFIC_BENEFITS, SURPLUS_IMAGES } from '../constants'
 import { getSurplusText, getTwitterShareUrl, getTwitterShareUrlForBenefit } from '../helpers'
