+++ apps/cowswap-frontend/src/modules/orderProgressBar/index.ts
@@ -1,5 +1,11 @@
+// Updaters
+
+export { ProgressBarExecutingOrdersUpdater } from './updaters/ProgressBarExecutingOrdersUpdater'
+
 // Containers
 export { SurplusModalSetup } from './containers/SurplusModalSetup'
 
 // Hooks
 export { useOrderSubmittedContent } from './hooks/useOrderSubmittedContent'
+
+export * from './types'
