+++ apps/cowswap-frontend/src/modules/orderProgressBar/progressBarStateMachine.md
@@ -2,7 +2,7 @@
 
 ```mermaid
 
-stateDiagram (former stateDiagram-v2)
+stateDiagram (former stateDiagramV2)
     state place_order <<fork>>
         [*] --> place_order: user place order
         place_order --> EOA
