+++ apps/cowswap-frontend/src/modules/orderProgressBar/types.ts
@@ -1,5 +1,10 @@
 import { SolverInfo } from '@cowprotocol/core'
-import type { CompetitionOrderStatus } from '@cowprotocol/cow-sdk'
+import { CompetitionOrderStatus, SupportedChainId } from '@cowprotocol/cow-sdk'
+import { Command } from '@cowprotocol/types'
+
+import type { Order } from 'legacy/state/orders/actions'
+
+import type { SurplusData } from 'common/hooks/useGetSurplusFiatValue'
 
 export type OrderProgressBarState = {
   countdown?: number | null
@@ -8,7 +13,6 @@ export type OrderProgressBarState = {
   solverCompetition?: CompetitionOrderStatus['value']
   progressBarStepName?: OrderProgressBarStepName
   previousStepName?: OrderProgressBarStepName
-  lastTimeChangedSteps?: number
   cancellationTriggered?: true
 }
 
@@ -24,3 +28,18 @@ export type OrderProgressBarStepName = happyPath | errorFlow | cancellationFlow
 type Unpacked<T> = T extends (infer U)[] ? U : never
 export type ApiSolverCompetition = Unpacked<CompetitionOrderStatus['value']>
 export type SolverCompetition = ApiSolverCompetition & Partial<SolverInfo>
+
+export type OrderProgressBarProps = {
+  stepName?: OrderProgressBarStepName
+  chainId: SupportedChainId
+  countdown?: number | null | undefined
+  solverCompetition?: SolverCompetition[]
+  totalSolvers?: number
+  order?: Order
+  debugMode?: boolean
+  showCancellationModal: Command | null
+  surplusData?: SurplusData
+  receiverEnsName?: string
+  navigateToNewOrder?: Command
+  isProgressBarSetup: boolean
+}
