--- apps/cowswap-frontend/src/modules/orderProgressBarMixed/orderProgressBar/index.ts
@@ -1,3 +1,0 @@
-export * from './useOrderProgressBarProps'
-export * from './types'
-export * from './ProgressBarExecutingOrdersUpdater'
